Pause stopwatch functionality
The plugin will now poll on page load until the stopwatch is
loaded and it will then pause it, so that wrong time is not logged
by mistake when closing or refreshing a Salesforce tab.

This is a client workaround and could be permanently fixed by
diabling the stopwatch in CaseTimer.js.

Authored-by: Ignacio Elizaga <ielizaga@pivotal.io>